+++ src/constants/index.js
@@ -9,3 +9,4 @@ export { default as resourceClasses } from './resourceClasses';
 export { default as resourceTypes } from './resourceTypes';
 export { default as validationEndPoint } from './validationEndPoint';
 export { default as defaultMclPageSize } from './defaultMclPageSize';
+export { default as hiddenAccordions } from './hiddenAccordions';