+++ src/settings/components/index.js
@@ -1,3 +1,4 @@
 export { default as MCLPaginationFields } from './MCLPaginationFields';
 export { default as SupplementaryPropertiesConfigForm } from './SupplementaryPropertiesConfigForm';
 export { default as SuppressFromDiscoveryFields } from './SuppressFromDiscoveryFields';
+export { default as HideAccordions } from './HideAccordions';