+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 8.1.0 In Progress
 * Upgrade `@folio/react-intl-safe-html` for `@folio/stripes` `v7` compatibility. ERM-1893.
 * Remove `OrgRoleSettings` from `src/settings`. ERM-1490
+* Add hide accordions option to settings. ERM-1762
 
 ## 8.0.1 2021-11-08
 * Handle 414 error exception when requesting an agreement line with large number of POLines. Refs ERM-1915